Bat out of the business
My own community is shutting down which explains alot why I haven't been commiting to this project at all.

Most likely I'm not going to contiune with any additions to this and I am willing to hand this project down to anyone else willing to.

I'm not sure when I'll return to the public FF2 scene but I'll still be here making content for TF2, just not right now with FF2.